+++ Colegio/src/main/java/com/Proyecto/Colegio/ColegioApplication.java
@@ -2,8 +2,10 @@ package com.Proyecto.Colegio;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"com.Proyecto.Colegio"})
 public class ColegioApplication {
 
 	public static void main(String[] args) {
