+++ Colegio/src/main/java/com/Proyecto/Colegio/dto/ColegioDto.java
@@ -1,0 +1,18 @@
+package com.Proyecto.Colegio.dto;
+
+import lombok.*;
+
+import java.util.List;
+
+@Getter
+@Setter
+@EqualsAndHashCode
+@ToString
+@AllArgsConstructor
+@NoArgsConstructor
+
+public class ColegioDto {
+    Long id;
+    String nombre;
+
+}
