+++ Colegio/src/main/java/com/Proyecto/Colegio/dto/CursoDto.java
@@ -1,0 +1,17 @@
+package com.Proyecto.Colegio.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode
+@AllArgsConstructor
+@NoArgsConstructor
+
+public class CursoDto {
+   Long id;
+   Integer grado;
+   String salon;
+   ColegioDto colegio;
+}
