+++ Colegio/src/main/java/com/Proyecto/Colegio/mapper/CursoMapper.java
@@ -1,13 +1,13 @@
 package com.Proyecto.Colegio.mapper;
 
-import com.Proyecto.Colegio.persistence.dto.CursoDto;
+import com.Proyecto.Colegio.dto.CursoDto;
 import com.Proyecto.Colegio.persistence.entity.CursoEntity;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 
 import java.util.List;
 
-@Mapper (componentModel = "spring", uses = {CursoMapper.class})
+@Mapper (componentModel = "spring", uses = {CursoMapper.class,ColegioMapper.class})
 public interface CursoMapper {
 
     CursoDto toDto(CursoEntity cursoEntity);
