+++ Colegio/src/main/java/com/Proyecto/Colegio/persistence/entity/CursoEntity.java
@@ -1,10 +1,7 @@
 package com.Proyecto.Colegio.persistence.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 import javax.persistence.*;
 import java.io.Serializable;
@@ -15,6 +12,8 @@ import java.io.Serializable;
 @Setter
 @EqualsAndHashCode
 @ToString
+@AllArgsConstructor
+@NoArgsConstructor
 public class CursoEntity implements Serializable {
 
     @Id
