+++ Colegio/src/main/java/com/Proyecto/Colegio/repository/CursoRepository.java
@@ -1,6 +1,5 @@
 package com.Proyecto.Colegio.repository;
 
-import com.Proyecto.Colegio.persistence.dto.CursoDto;
 import com.Proyecto.Colegio.persistence.entity.CursoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,8 +10,11 @@ import java.util.List;
 @Repository
 public interface CursoRepository extends JpaRepository <CursoEntity,Long> {
 
-    @Query(value = "SELECT new com.Proyecto.Colegio.persistence.dto.CursoDto(C.id,C.grado,C.salon)" +
-            "FROM CursoEntity C WHERE C.colegio.id = :id")
-    List<CursoDto> getAllCurses(@Param("id") Long id);
+//    @Query(value = "SELECT new com.Proyecto.Colegio.dto.CursoDto(C.id,C.grado,C.salon)" +
+//            "FROM CursoEntity C WHERE C.colegio.id = :id")
+//    List<CursoDto> getAllCurses(@Param("id") Long id);
+
+    @Query(value = "SELECT C.grado,C.salon FROM CursoEntity C WHERE C.colegio.id = :id")
+    List<CursoEntity> getAllCurses(@Param("id") Long id);
 
 }
