+++ Colegio/src/main/java/com/Proyecto/Colegio/service/ColegioService.java
@@ -1,11 +1,6 @@
 package com.Proyecto.Colegio.service;
 
-import com.Proyecto.Colegio.persistence.dto.CursoDto;
-import com.Proyecto.Colegio.repository.ColegioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 public class ColegioService {
